buffer_cache: Strip down and swap to sparse resident memory management
DMA doesn't work, will fix it later

+++ src/video_core/texture_cache/tile_manager.h
@@ -25,9 +25,9 @@ public:
     ~TileManager();
 
     void TileImage(Image& in_image, std::span<vk::BufferImageCopy> buffer_copies,
-                   vk::Buffer out_buffer, u32 out_offset, u32 copy_size);
+                   vk::Buffer out_buffer, u64 out_offset, u32 copy_size);
 
-    Result DetileImage(vk::Buffer in_buffer, u32 in_offset, const ImageInfo& info);
+    Result DetileImage(vk::Buffer in_buffer, u64 in_offset, const ImageInfo& info);
 
 private:
     vk::Pipeline GetTilingPipeline(const ImageInfo& info, bool is_tiler);
